+++ litho-testing/src/main/java/com/facebook/litho/testing/sections/TestSectionCreator.java
@@ -115,7 +115,8 @@ public class TestSectionCreator {
         long uptimeMillis,
         int firstVisibleIndex,
         int lastVisibleIndex,
-        ChangesInfo changesInfo) {
+        ChangesInfo changesInfo,
+        int globalOffset) {
       onDataRendered = true;
       mChangesInfo = changesInfo;
     }